Combine errors into a single module
