--- lib/poms/errors/authentication_error.rb
@@ -1,12 +1,0 @@
-module Poms
-  module Errors
-    # Custom Error class to signal missing authorization attributes
-    class AuthenticationError < StandardError
-      attr_reader :message
-
-      def initialize(message)
-        @message = message
-      end
-    end
-  end
-end
