--- lib/poms/errors/http_error.rb
@@ -1,7 +1,0 @@
-module Poms
-  module Errors
-    # Base exception for all Poms-specific HTTP errors.
-    class HttpError < StandardError
-    end
-  end
-end
