--- lib/poms/errors/http_missing_error.rb
@@ -1,9 +1,0 @@
-require 'poms/errors/http_error'
-
-module Poms
-  module Errors
-    # Wrapper exception for dealing with driver-agnostic 404 responses.
-    class HttpMissingError < HttpError
-    end
-  end
-end
