--- lib/poms/errors/http_server_error.rb
@@ -1,9 +1,0 @@
-require 'poms/errors/http_error'
-
-module Poms
-  module Errors
-    # Wrapper exception for dealing with driver-agnostic 500 responses.
-    class HttpServerError < HttpError
-    end
-  end
-end
